Mrtk v next input and utilities debug clean up (#1975)
* Cleaning up Debug references to a utility class
Enables clearer visibility for what is being debugged and an easier path to exclude them from a build.

* Added license text to new classes
Assigned the EventSystem classes to the new .Events namespace

* Cleaned up, separated and documented the Extension methods.
Still needs a second pass for individual methods / properties.

* Defines removed around Assert calls in the Utility class.

* Removed Debug.LogErrorFormat / Debug.LogWarningFormat calls and replaced with String Interpolation
