+++ Assets/MixedRealityToolkit/Boundary/MixedRealityBoundarySystem.cs
@@ -11,6 +11,8 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
     /// </summary>
     public class MixedRealityBoundaryManager : BaseManager, IMixedRealityBoundarySystem
     {
+        IMixedRealityInputSystem inputSystem;
+        
         /// <summary>
         /// MixedRealityBoundaryManager constructor
         /// </summary>
@@ -26,6 +28,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         public override void Initialize()
         {
             // TODO Initialize stuff 
+            inputSystem = Internal.Managers.MixedRealityManager.Instance.GetManager<IMixedRealityInputSystem>();
         }
 
         /// <summary>
