+++ Assets/MixedRealityToolkit/EventSystem/GenericBaseEventData.cs
@@ -1,6 +1,9 @@
-﻿using UnityEngine.EventSystems;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Microsoft.MixedReality.Toolkit.Internal
+using UnityEngine.EventSystems;
+
+namespace Microsoft.MixedReality.Toolkit.Internal.Events
 {
     public class GenericBaseEventData : BaseEventData
     {
