+++ Assets/MixedRealityToolkit/EventSystem/IEventSource.cs
@@ -1,6 +1,9 @@
-﻿using System.Collections;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Microsoft.MixedReality.Toolkit.Internal
+using System.Collections;
+
+namespace Microsoft.MixedReality.Toolkit.Internal.Events
 {
     public interface IEventSource : IEqualityComparer
     {
