--- Assets/MixedRealityToolkit/InputSystem/EventData.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4d802166a8d2f6e4c978763957e39a1d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
