+++ Assets/MixedRealityToolkit/InputSystem/Focus/RayStep.cs
@@ -58,8 +58,8 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem.Focus
         /// <returns></returns>
         public static Vector3 GetPointByDistance(RayStep[] steps, float distance)
         {
-            Debug.Assert(steps != null);
-            Debug.Assert(steps.Length > 0);
+             Debug.Assert(steps != null);
+             Debug.Assert(steps.Length > 0);
 
             Vector3 point = Vector3.zero;
             float remainingDistance = distance;
@@ -95,8 +95,8 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem.Focus
         /// <returns></returns>
         public static RayStep GetStepByDistance(RayStep[] steps, float distance)
         {
-            Debug.Assert(steps != null);
-            Debug.Assert(steps.Length > 0);
+             Debug.Assert(steps != null);
+             Debug.Assert(steps.Length > 0);
 
             RayStep step = new RayStep();
             float remainingDistance = distance;
@@ -132,8 +132,8 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem.Focus
         /// <returns></returns>
         public static Vector3 GetDirectionByDistance(RayStep[] steps, float distance)
         {
-            Debug.Assert(steps != null);
-            Debug.Assert(steps.Length > 0);
+             Debug.Assert(steps != null);
+             Debug.Assert(steps.Length > 0);
 
             return GetStepByDistance(steps, distance).Direction;
         }
