+++ Assets/MixedRealityToolkit/InputSystem/Sources/InteractionInputSources.cs
@@ -603,7 +603,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem.Sources
         {
             if (capability.IsAvailable)
             {
-                Debug.Assert(capability.IsSupported);
+                 Debug.Assert(capability.IsSupported);
 
                 reading = capability.CurrentReading;
                 return true;
@@ -731,7 +731,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem.Sources
 
             foreach (var pointerOption in pointerOptions)
             {
-                Debug.Assert(pointerOption.TargetController != Handedness.None, "Interaction Source Pointer must be set to Left, Right, or Both.");
+                 Debug.Assert(pointerOption.TargetController != Handedness.None, "Interaction Source Pointer must be set to Left, Right, or Both.");
 
                 if (interactionSource.handedness == InteractionSourceHandedness.Unknown ||
                     interactionSource.handedness == InteractionSourceHandedness.Left && pointerOption.TargetController == Handedness.Right ||
@@ -789,8 +789,8 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem.Sources
         /// <param name="sourceData">GenericInputPointingSource structure to update.</param>
         private static void UpdateInteractionSource(InteractionSourceState interactionSourceState, InteractionInputSource sourceData)
         {
-            Debug.Assert(interactionSourceState.source.id == sourceData.Source.id, "An UpdateSourceState call happened with mismatched source ID.");
-            Debug.Assert(interactionSourceState.source.kind == sourceData.Source.kind, "An UpdateSourceState call happened with mismatched source kind.");
+             Debug.Assert(interactionSourceState.source.id == sourceData.Source.id, "An UpdateSourceState call happened with mismatched source ID.");
+             Debug.Assert(interactionSourceState.source.kind == sourceData.Source.kind, "An UpdateSourceState call happened with mismatched source kind.");
 
             Vector3 newPointerPosition;
             sourceData.PointerPosition.IsAvailable = interactionSourceState.sourcePose.TryGetPosition(out newPointerPosition, InteractionSourceNode.Pointer);
