+++ Assets/MixedRealityToolkit/InputSystem/Utilities/Interactions/KeywordAndResponse.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
 using UnityEngine;
 using UnityEngine.Events;
