+++ Assets/MixedRealityToolkit/_Core/Extensions/DoubleExtensions.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 907120ad70f1b9e468d249db24f0b1e5
-timeCreated: 1473730751
-licenseType: Pro
+guid: c32903a3958f2e047a2f3fa33c5056f4
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
