+++ Assets/MixedRealityToolkit/_Core/Extensions/RayExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9c4063f81a7575489c75c325f628075
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
