--- Assets/MixedRealityToolkit/_Core/Plugins.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5d2d4caad575b8c45861c305e69f2ad9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
